+++ db/src/db/DB2024Team13_main.java
@@ -4,7 +4,7 @@ import db.view.DB2024Team13_loginWindow;
 
 public class DB2024Team13_main {
     public static void main(String[] args) {
-        // 로그인 창 표시
+        // 로그인 창으로 시작
         DB2024Team13_loginWindow.displayLoginWindow();
     }
 }
