+++ db/src/db/view/DB2024Team13_addRestaurantWindow.java
@@ -1,8 +1,9 @@
 package db.view;
 
-import db.controller.DB2024Team13_connection;
-
 import javax.swing.*;
+
+import db.model.DB2024Team13_connection;
+
 import java.awt.*;
 import java.sql.*;
 import java.util.ArrayList;
